Implement q_sort
Implement a function to sort the queue node
in ascending or descending order, which determined by
parameter descend.
The function do nothing if the queue is NULL,
empty or singular.

The sorting functionality of this function is
implemented using merge_sort.

To achieve the effect of merge_sort,
I have additionally implemented __merge to
merge two linked lists.
This function takes l and r as input,
representing two linked lists, and descend,
which determines the sorting order.

Additionally, I have implemented __merge_sort to
recursively split the linked list and
call __merge for merging.
This function takes l as input,
representing the head of the linked list,
and descend to determine the sorting order.
If l is NULL or l->next is NULL,
it directly returns l.
Change-Id: I686d6b913a12d44568cec362706d77ffbe6f370c