Implement q_merge
Implement a function that merges multiple
sorted queues into the first queue.

The function returns the length of the merged queue.
If head is NULL or chain is empty, it directly returns 0.

Currently, the merging is performed
using a pairwise merging approach,
which generally results in more balanced lengths
and is faster than merging one by one in most cases.

Change-Id: Ie0b2ceeada0c43b8d30845b255eb5b899897ea8e