Rewrite q_merge
The previous implementation required allocating
an array of size 1000 on the stack to store
multiple queues. This approach had two major drawbacks:

It consumed a significant amount of stack space.
It could fail when the number of queues exceeded 1000.

The new implementation is inspired by the Linux kernel.
It uses the next pointer of list_head to maintain
 a singly linked list, and leverages the list_sort merge
function to handle the merging process.

Compared to the previous approach,
which relied on storing comparison results via indirect pointers,
 this method is nearly twice as fast in terms of performance.

During the merge process, the prev pointer of list_head is
repurposed to maintain the ordering between queues.
Two list_head objects, r0 and r1,
are used to hold the queues pending merge and
the result of merged queues, respectively.
The algorithm repeatedly performs merging and
swaps r0 and r1 until r0 contains only one queue
 and r1 is empty — at which point the merging is complete.

While both the old and new methods use pairwise merging to
 ensure efficient combining of queues, the new implementation
can theoretically handle an arbitrary number of queues
without consuming additional memory as the number increases.

Change-Id: If3aa2eab93d40dec3ee44fd579cd7080638e89f3